// frontend/src/api/model/workspaceDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * TeamPrompts.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface WorkspaceDto {
  id: string;
  name: string;
  /** @nullable */
  key: string | null;
  /** @nullable */
  avatarUrl: string | null;
  sortOrder: number;
  isSystem: boolean;
  scriptCount: number;
  promptCount: number;
  createdAt: string;
  updatedAt: string;
}